Rename deployer.py to deployer.py
